+++ Dimension/TemperatureDimension.h
@@ -1,72 +1,16 @@
 #ifndef STATIC_DIMENSION_TEMPERATURE_H
 #define STATIC_DIMENSION_TEMPERATURE_H
 
-#include "BaseDimension.h"
+#include "Dimension_Impl/FundamentalDimensions/TemperatureDimension_Impl.h"
 
 namespace Dimension
 {
-   struct TemperatureType {};
-   struct Celsius;
-
-   template<typename TemperatureUnit>
-   concept IsTemperatureUnit = std::is_same_v<typename TemperatureUnit::Dim, TemperatureType>;
-
-   template<typename Unit>
-   struct TemperatureUnit : public BaseUnit<Unit>
-   {
-   public:
-      using BaseUnit<Unit>::BaseUnit;
-
-      using Dim = TemperatureType;
-      using Primary = Celsius;
-   };
 
    struct Celsius : public TemperatureUnit<Celsius> { public: using TemperatureUnit::TemperatureUnit; };
    struct Fahrenheit : public TemperatureUnit<Fahrenheit> { public: using TemperatureUnit::TemperatureUnit; };
    struct Kelvin : public TemperatureUnit<Kelvin> { public: using TemperatureUnit::TemperatureUnit; };
    struct Rankine : public TemperatureUnit<Rankine> { public: using TemperatureUnit::TemperatureUnit; };
 
-   template<typename T>
-   struct is_temperature : std::is_convertible<T, BaseDimension<std::tuple<Celsius>, std::tuple<>>> {};
-
-   template<typename T>
-   constexpr bool is_temperature_v = is_temperature<T>::value;
-
-   template<typename T>
-   concept temperature_type = is_temperature_v<T>;
-
-   template<IsTemperatureUnit T>
-   PrecisionType getTemperature(temperature_type auto obj)
-   {
-      return obj.template GetVal<std::tuple<T>, std::tuple<>>();
-   }
-
-   template<IsTemperatureUnit Unit>
-   class Temperature : public BaseDimension<std::tuple<Unit>, std::tuple<>>
-   {
-   public:
-      using BaseDimension<std::tuple<Unit>, std::tuple<>>::BaseDimension;
-
-      Temperature() : BaseDimension<std::tuple<Unit>, std::tuple<>>::BaseDimension(0.0) {}
-
-      Temperature(double val) : BaseDimension<std::tuple<Unit>, std::tuple<>>::BaseDimension(val) {}
-
-      template<IsTemperatureUnit T>
-      Temperature(const BaseDimension<std::tuple<T>, std::tuple<>>& base) : BaseDimension<std::tuple<Unit>, std::tuple<>>::BaseDimension(base.template GetVal<std::tuple<Unit>, std::tuple<>>()) {}
-
-      template<IsTemperatureUnit T>
-      [[deprecated("Use the free function getTemperature() instead.")]]
-      double GetTemperature() const
-      {
-         return getTemperature<T>(*this);
-      }
-   };
-
-   template<IsTemperatureUnit T>
-   Temperature(T) -> Temperature<T>;
-
-   template<IsTemperatureUnit TemperatureUnit>
-   Temperature(BaseDimension<std::tuple<TemperatureUnit>, std::tuple<>>) -> Temperature<TemperatureUnit>;
 
    template<> struct Conversion<Celsius, Fahrenheit>
    {